Update SK19Lift.java
SK19 was updated to include an ArrayList to work as a lookup table. This needs to be changed to be a 2D static array that is intiated at the class level not in the constructor and structured to look like a matrix. The 2D array needs to be of doubles and do a type conversion of doubles to boolean  using 1.0 as false and 2.0 as true.