Create new bringup branch.
This branch is intended for use during robot bringup and class debug. As it stands, it supports only the base drivetrain and gear shift. Once these components are debugged, we will merge new subsystems from the develop branch one at a time and debug each. When all subsystems are working, we will merge this back to develop and continue using that branch as before.