Create CargoIntakeSystem.java
Created CargoIntakeSystem that runs the cargo system intake system automatically and allows the cargo once it gets into the octopus to stop everything and pull it back in. THERE MAY BE AN ISSUE WITH GETTING CARGO IN, the cargo may not be able to get past the intake system stopping it from get into the robot, meaning that we may need to add a second sensor to the intake mechanism allowing it to see we have a piece of cargo and we are ready to finish pulling it in.