Cleaned up code and added spots to put useful software
New folder has been created to put useful firmware and any other things that require to be loaded onto individuals laptops. The code has now been cleaned up to be much more readable and understandable. We are still going to need to figure out if we need new sensors and other things to be really ready for Greenville and they'll need to be fitted on Saturday to allow testing on that day as well as on Wednesday. Code needs to be tested but it should be working as all new code has not been implemented yet and needs to be implemented slowly.