New Climb Changes
Ports.Java has been updated to include the newly required solenoids and motors. AT THIS POINT THERE ARE NO MORE FREE SOLENOIDS. Everything will need to be done via motors and the climb code now needs to be updated to reflect the use of two motors along with the use of pneumatics.