Fixed Cargo Intake commands
The cargo intake commands have been fixed to allow the operator to push a button and continue the second part of the intake to pull things in. This has not been mapped to an OI button as we need to figure where it's going to go, and we may want to do the same arm check and pull down to guarantee that we can pull in a piece of cargo.